requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
numpy
scikit-learn
requests
python-multipart
jinja2
redis
asyncio